SubscriptionManager: move push subscription gen logic to Core Director
Motivation: cleans up the subscription manager a bit.

In order to successfully create a user, we must have at least 1 subscription in the core module.

We should also hydrate the user immediately after success, previously not being done.
